Made an Base_Agent abstract class and have other Agent classes inherit from it and override different methods. Separate agents into DQN_Agent, Simple_Q_Learning_Agent, Random_Walking_Agent
